修复图灵机器人在群聊中回复异常
修复图灵机器人在群聊中回复异常
